adicionada a classe Campanha com o pacote correto
